Add min-idle time for server.
the autoscaler will shutdown any server that is not busy but has reached the min_age parameter (30m currently)

This feature adds a new flag that basically resets the age to 0 every time a new build is executed.

This will allow us to decrease the min-age but allow the servers to be used for a longer period of time during busy hours, thus avoiding starting up too many servers and decreasing the build startup times

This is a feature added by a contributor to the original autoscaler code. harness is very slow to accept contributions so the PR has sat there for about 1 month. https://github.com/drone/autoscaler/pull/125